Add cache reader and cache converter
This adds new cahce reader and cache converter types to the informers
package.  The cache reader provides a read-only view across all of the
caches backing a cross-namespace informer.  The cache converter wraps
any cache.Indexer that holds unstructured types and converts them to
concrete types as they are read and written.