Send delete events on namespace removal
In order to match the behavior of MultiListWatcher, we need to send
delete events for everything in the cache for a namespace when that
namespace is removed from the set of watched namespaces.  This does
that in the informer's RemoveNamespace method by simply listing
everything in the store after stopping the informer, and manually
invoking each handler's OnDelete method for each object.